requirements.txt
pillow
tqdm
open-tamil